code clean up

+++ src/compiler/scanner.h
@@ -417,7 +417,7 @@ class Scanner {
   // warning message
   void ProcessWarning() {
     wcout << GetToken()->GetFileName() << ":" << GetToken()->GetLineNumber() + 1
-	  << ": Parse warning: Unknown token: '" << (int)cur_char << "'" << endl;
+	  << ": Parse warning: Unknown token: '" << cur_char << "'" << endl;
   }
 
   /****************************
